Add build flags to support OSX and windows
